Refactor TodoList to TaskList to align with other classes
